requirements_dev.txt
# Generic python dependencies
pip==26.0
pre-commit==4.5.1
ruff==0.14.14
ty==0.0.14
voluptuous-stubs==0.1.1

# Home Assistant testing dependencies
pytest-homeassistant-custom-component==0.13.310
syrupy>=4.6.1

# Component dependencies
hassil>=1.7.0
home-assistant-intents>=2024.4.3
